Offline Dictionary Lookup
In SampleCode\WordGrid\WorkerRole1\WorkerRole.fs around line 102 I made
some modifications that help with offline debugging. The file path for
OpenRead will need to be changed depending on where you have stored the
dictionary.txt file.  I have added the one I used to
SampleCode\WordGrid\WorkerRole1\dictionary.txt.  This is very helpful
for experimenting/debugging with the code locally but will need to be
changed when migrated to the cloud